Splitted the project into two parts: one responsible for the building (handling of the defined DSL) and one responsible for the managing of the built objects in an EntityManager. New tests for the builder part implemented.
